servers/fastapi-mcp-server/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
psycopg2-binary>=2.9.9
fastmcp>=2.0.0
pydantic